Title: 자연수 n개의 합, Time: 96ms, Memory: 8MB, Status: Passed - CodeTree

+++ 241031/자연수 n개의 합/sum-of-n-natural-numbers.java
@@ -27,7 +27,7 @@ public class Main {
     public static void main(String[] args) throws Exception {
         // 여기에 코드를 작성해주세요.
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        s = Integer.parseInt(br.readLine());
+        s = Long.parseLong(br.readLine());
 
         long ans = binarySearch();
